+++ tests/clients/test_alerts.py
@@ -39,13 +39,13 @@ class TestAlertClient(object):
         alert_client.search(query)
         assert mock_session.post.call_args[0][0] == u"/svc/api/v1/query-alerts"
 
-    def test_get_query_details_when_not_given_tenant_id_posts_expected_data(
+    def test_get_details_when_not_given_tenant_id_posts_expected_data(
         self, mock_session, user_context, successful_response
     ):
         mock_session.post.return_value = successful_response
         alert_client = AlertClient(mock_session, user_context)
         alert_ids = ["ALERT_ID_1", "ALERT_ID_2"]
-        alert_client.get_query_details(alert_ids)
+        alert_client.get_details(alert_ids)
         post_data = json.loads(mock_session.post.call_args[1]["data"])
         assert (
             post_data["tenantId"] == TENANT_ID_FROM_RESPONSE
@@ -53,23 +53,23 @@ class TestAlertClient(object):
             and post_data["alertIds"][1] == "ALERT_ID_2"
         )
 
-    def test_get_query_details_when_given_single_alert_id_posts_expected_data(
+    def test_get_details_when_given_single_alert_id_posts_expected_data(
         self, mock_session, user_context, successful_post
     ):
         alert_client = AlertClient(mock_session, user_context)
-        alert_client.get_query_details("ALERT_ID_1")
+        alert_client.get_details("ALERT_ID_1")
         post_data = json.loads(mock_session.post.call_args[1]["data"])
         assert (
             post_data["tenantId"] == TENANT_ID_FROM_RESPONSE
             and post_data["alertIds"][0] == "ALERT_ID_1"
         )
 
-    def test_get_query_details_when_given_tenant_id_posts_expected_data(
+    def test_get_details_when_given_tenant_id_posts_expected_data(
         self, mock_session, user_context, successful_post
     ):
         alert_client = AlertClient(mock_session, user_context)
         alert_ids = ["ALERT_ID_1", "ALERT_ID_2"]
-        alert_client.get_query_details(alert_ids, "some-tenant-id")
+        alert_client.get_details(alert_ids, "some-tenant-id")
         post_data = json.loads(mock_session.post.call_args[1]["data"])
         assert (
             post_data["tenantId"] == "some-tenant-id"
@@ -77,12 +77,12 @@ class TestAlertClient(object):
             and post_data["alertIds"][1] == "ALERT_ID_2"
         )
 
-    def test_get_query_details_posts_to_expected_url(
+    def test_get_details_posts_to_expected_url(
         self, mock_session, user_context, successful_post
     ):
         alert_client = AlertClient(mock_session, user_context)
         alert_ids = ["ALERT_ID_1", "ALERT_ID_2"]
-        alert_client.get_query_details(alert_ids)
+        alert_client.get_details(alert_ids)
         assert mock_session.post.call_args[0][0] == "/svc/api/v1/query-details"
 
     def test_resolve_when_not_given_tenant_id_posts_expected_data(